requirements.txt
implicit==0.4.2
joblib==0.16.0
numpy==1.19.1
pandas==1.1.1
pyaml==20.4.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.21.3
scikit-optimize==0.7.4
scipy==1.3.1
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.48.2
